Spread item tests over dedicated files
