+++ Tests/PlayerTests/ItemRemovalTests.swift
@@ -1,0 +1,93 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+@testable import Player
+
+import AVFoundation
+import Circumspect
+import Nimble
+import XCTest
+
+@MainActor
+final class ItemRemovalTests: XCTestCase {
+    func testRemovePreviousItem() {
+        let item1 = AVPlayerItem(url: URL(string: "https://www.server.com/item1.m3u8")!)
+        let item2 = AVPlayerItem(url: URL(string: "https://www.server.com/item2.m3u8")!)
+        let item3 = AVPlayerItem(url: URL(string: "https://www.server.com/item3.m3u8")!)
+        let player = DequePlayer(items: [item1, item2, item3])
+        player.advanceToNextItem()
+        expect(player.currentItem).to(equal(item2))
+
+        player.remove(item1)
+        expect(player.items()).to(equalDiff([item2, item3]))
+        expect(player.nextItems()).to(equalDiff([item3]))
+        expect(player.previousItems()).to(beEmpty())
+    }
+
+    func testRemoveCurrentItem() {
+        let item1 = AVPlayerItem(url: URL(string: "https://www.server.com/item1.m3u8")!)
+        let item2 = AVPlayerItem(url: URL(string: "https://www.server.com/item2.m3u8")!)
+        let item3 = AVPlayerItem(url: URL(string: "https://www.server.com/item3.m3u8")!)
+        let player = DequePlayer(items: [item1, item2, item3])
+        player.advanceToNextItem()
+        expect(player.currentItem).to(equal(item2))
+
+        player.remove(item2)
+        expect(player.currentItem).to(equal(item3))
+        expect(player.items()).to(equalDiff([item1, item3]))
+        expect(player.nextItems()).to(beEmpty())
+        expect(player.previousItems()).to(equalDiff([item1]))
+    }
+
+    func testRemoveCurrentLastItem() {
+        let item1 = AVPlayerItem(url: URL(string: "https://www.server.com/item1.m3u8")!)
+        let item2 = AVPlayerItem(url: URL(string: "https://www.server.com/item2.m3u8")!)
+        let player = DequePlayer(items: [item1, item2])
+        player.advanceToNextItem()
+        expect(player.currentItem).to(equal(item2))
+
+        player.remove(item2)
+        expect(player.currentItem).to(beNil())
+        expect(player.items()).to(equalDiff([item1]))
+        expect(player.nextItems()).to(beEmpty())
+        expect(player.previousItems()).to(equalDiff([item1]))
+    }
+
+    func testRemoveNextItem() {
+        let item1 = AVPlayerItem(url: URL(string: "https://www.server.com/item1.m3u8")!)
+        let item2 = AVPlayerItem(url: URL(string: "https://www.server.com/item2.m3u8")!)
+        let item3 = AVPlayerItem(url: URL(string: "https://www.server.com/item3.m3u8")!)
+        let player = DequePlayer(items: [item1, item2, item3])
+        player.advanceToNextItem()
+        expect(player.currentItem).to(equal(item2))
+
+        player.remove(item3)
+        expect(player.items()).to(equalDiff([item1, item2]))
+        expect(player.nextItems()).to(beEmpty())
+        expect(player.previousItems()).to(equalDiff([item1]))
+    }
+
+    func testRemoveForeignItem() {
+        let item = AVPlayerItem(url: URL(string: "https://www.server.com/item.m3u8")!)
+        let foreignItem = AVPlayerItem(url: URL(string: "https://www.server.com/foreign.m3u8")!)
+        let player = DequePlayer(items: [item])
+        player.remove(foreignItem)
+        expect(player.items()).to(equalDiff([item]))
+    }
+
+    func testRemoveAllItems() {
+        let item1 = AVPlayerItem(url: URL(string: "https://www.server.com/item1.m3u8")!)
+        let item2 = AVPlayerItem(url: URL(string: "https://www.server.com/item2.m3u8")!)
+        let item3 = AVPlayerItem(url: URL(string: "https://www.server.com/item3.m3u8")!)
+        let player = DequePlayer(items: [item1, item2, item3])
+        player.advanceToNextItem()
+        expect(player.currentItem).to(equal(item2))
+
+        player.removeAllItems()
+        expect(player.items()).to(beEmpty())
+        expect(player.currentItem).to(beNil())
+    }
+}
