+++ Tests/PlayerTests/PlayerTests.swift
@@ -11,7 +11,15 @@ import Nimble
 import XCTest
 
 @MainActor
-final class PlayerDeallocationTests: XCTestCase {
+final class PlayerTests: XCTestCase {
+    func testEmpty() {
+        let player = DequePlayer()
+        expect(player.currentItem).to(beNil())
+        expect(player.items()).to(beEmpty())
+        expect(player.nextItems()).to(beEmpty())
+        expect(player.previousItems()).to(beEmpty())
+    }
+
     func testPlayerDeallocation() {
         let item = AVPlayerItem(url: Stream.onDemand.url)
         var player: Player? = Player(item: item)
